+++ tests/broker_test.py
@@ -22,6 +22,30 @@ class ShutdownTest(testlib.TestCase):
         actual_close()
 
 
+class DeferTest(testlib.TestCase):
+    klass = mitogen.core.Broker
+
+    def test_defer(self):
+        latch = mitogen.core.Latch()
+        broker = self.klass()
+        try:
+            broker.defer(lambda: latch.put(123))
+            self.assertEquals(123, latch.get())
+        finally:
+            broker.shutdown()
+            broker.join()
+
+    def test_defer_after_shutdown(self):
+        latch = mitogen.core.Latch()
+        broker = self.klass()
+        broker.shutdown()
+        broker.join()
+
+        e = self.assertRaises(mitogen.core.Error,
+            lambda: broker.defer(lambda: latch.put(123)))
+        self.assertEquals(e.args[0], mitogen.core.Waker.dead_msg)
+
+
 class DeferSyncTest(testlib.TestCase):
     klass = mitogen.core.Broker